pyproject.toml
[tool.poetry]
name = "llamaindex-rag"
version = "0.1.0"
description = ""
authors = ["DY Lim <dlagus6412@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
aiohappyeyeballs = "2.4.3"
aiohttp = "3.10.10"
aiosignal = "1.3.1"
altair = "5.4.1"
annotated-types = "0.7.0"
anthropic = "0.39.0"
anyio = "4.6.2.post1"
appdirs = "1.4.4"
appnope = "0.1.4"
asttokens = "2.4.1"
attrs = "24.2.0"
backcall = "0.2.0"
beautifulsoup4 = "4.12.3"
bleach = "6.2.0"
blinker = "1.8.2"
bm25s = "0.2.3"
cachetools = "5.5.0"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
click = "8.1.7"
dataclasses-json = "0.6.7"
datasets = "3.1.0"
decorator = "5.1.1"
defusedxml = "0.7.1"
deprecated = "1.2.14"
dill = "0.3.8"
dirtyjson = "1.0.8"
distro = "1.9.0"
docopt = "0.6.2"
docstring-parser = "0.16"
executing = "2.1.0"
fastapi = "0.115.4"
fastjsonschema = "2.20.0"
filelock = "3.16.1"
frozenlist = "1.5.0"
fsspec = "2024.9.0"
gitdb = "4.0.11"
gitpython = "3.1.43"
google-api-core = "2.22.0"
google-auth = "2.36.0"
google-cloud-aiplatform = "1.71.1"
google-cloud-appengine-logging = "1.5.0"
google-cloud-audit-log = "0.3.0"
google-cloud-bigquery = "3.26.0"
google-cloud-core = "2.4.1"
google-cloud-documentai = "3.0.0"
google-cloud-firestore = "2.19.0"
google-cloud-logging = "3.11.3"
google-cloud-resource-manager = "1.13.0"
google-cloud-storage = "2.18.2"
google-crc32c = "1.6.0"
google-resumable-media = "2.7.2"
googleapis-common-protos = "1.65.0"
greenlet = "3.1.1"
grpc-google-iam-v1 = "0.13.1"
grpcio = "1.67.1"
grpcio-status = "1.67.1"
h11 = "0.14.0"
httpcore = "1.0.6"
httpx = "0.27.2"
httpx-sse = "0.4.0"
huggingface-hub = "0.26.2"
idna = "3.10"
importlib-metadata = "8.5.0"
ipython = "8.12.3"
jedi = "0.19.1"
jinja2 = "3.1.4"
jiter = "0.7.0"
joblib = "1.4.2"
jsonpatch = "1.33"
jsonpointer = "3.0.0"
jsonschema = "4.23.0"
jsonschema-specifications = "2024.10.1"
jupyter-client = "8.6.3"
jupyter-core = "5.7.2"
jupyterlab-pygments = "0.3.0"
langchain = "0.3.7"
langchain-community = "0.3.5"
langchain-core = "0.3.15"
langchain-google-vertexai = "2.0.7"
langchain-openai = "0.2.6"
langchain-text-splitters = "0.3.2"
langsmith = "0.1.140"
llama-cloud = "0.1.4"
llama-index = "0.11.22"
llama-index-agent-openai = "0.3.4"
llama-index-cli = "0.3.1"
llama-index-core = "0.11.22"
llama-index-embeddings-openai = "0.2.5"
llama-index-embeddings-vertex = "0.2.1"
llama-index-indices-managed-llama-cloud = "0.4.0"
llama-index-legacy = "0.9.48.post3"
llama-index-llms-openai = "0.2.16"
llama-index-llms-vertex = "0.3.7"
llama-index-multi-modal-llms-openai = "0.2.3"
llama-index-program-openai = "0.2.0"
llama-index-question-gen-openai = "0.2.0"
llama-index-readers-file = "0.2.2"
llama-index-readers-llama-parse = "0.3.0"
llama-index-retrievers-bm25 = "0.4.0"
llama-index-storage-docstore-firestore = "0.2.0"
llama-index-storage-kvstore-firestore = "0.3.0"
llama-index-vector-stores-vertexaivectorsearch = "0.1.0"
llama-parse = "0.5.13"
markdown-it-py = "3.0.0"
markupsafe = "3.0.2"
marshmallow = "3.23.1"
matplotlib-inline = "0.1.7"
mdurl = "0.1.2"
mistune = "3.0.2"
multidict = "6.1.0"
multiprocess = "0.70.16"
mypy-extensions = "1.0.0"
narwhals = "1.13.2"
nbclient = "0.10.0"
nbconvert = "7.16.4"
nbformat = "5.10.4"
nest-asyncio = "1.6.0"
networkx = "3.4.2"
nltk = "3.9.1"
numpy = "1.26.4"
openai = "1.54.2"
opentelemetry-api = "1.28.0"
orjson = "3.10.11"
packaging = "24.1"
pandas = "2.2.3"
pandocfilters = "1.5.1"
parso = "0.8.4"
pexpect = "4.9.0"
pickleshare = "0.7.5"
pillow = "10.4.0"
pipreqs = "0.5.0"
platformdirs = "4.3.6"
prompt-toolkit = "3.0.48"
propcache = "0.2.0"
proto-plus = "1.25.0"
protobuf = "5.28.3"
ptyprocess = "0.7.0"
pure-eval = "0.2.3"
pyarrow = "15.0.2"
pyasn1 = "0.6.1"
pyasn1-modules = "0.4.1"
pydantic = "2.9.2"
pydantic-settings = "2.6.1"
pydantic-core = "2.23.4"
pydeck = "0.9.1"
pygments = "2.18.0"
pypdf = "4.3.1"
pysbd = "0.3.4"
pystemmer = "2.2.0.3"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2024.2"
pyyaml = "6.0.2"
pyzmq = "26.2.0"
ragas = "0.2.3"
referencing = "0.35.1"
regex = "2024.9.11"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
rich = "13.9.4"
rpds-py = "0.21.0"
rsa = "4.9"
scipy = "1.14.1"
shapely = "2.0.6"
six = "1.16.0"
smmap = "5.0.1"
sniffio = "1.3.1"
soupsieve = "2.6"
sqlalchemy = "2.0.35"
stack-data = "0.6.3"
starlette = "0.41.2"
streamlit = "1.39.0"
striprtf = "0.0.26"
tenacity = "8.5.0"
tiktoken = "0.8.0"
tinycss2 = "1.4.0"
toml = "0.10.2"
tornado = "6.4.1"
tqdm = "4.67.0"
traitlets = "5.14.3"
typing-inspect = "0.9.0"
typing-extensions = "4.12.2"
tzdata = "2024.2"
urllib3 = "2.2.3"
uvicorn = "0.32.0"
wcwidth = "0.2.13"
webencodings = "0.5.1"
wrapt = "1.16.0"
xxhash = "3.5.0"
yarg = "0.1.9"
yarl = "1.17.1"
zipp = "3.20.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
